Tighten mobile shell actions
