+++ src/app/shell/MobileTabBar.tsx
@@ -1,5 +1,6 @@
-import { Home, MessageSquare, Plus, LayoutGrid, Bot, Users, BookOpen, FileText, Link, Sparkles, Settings, User } from "lucide-react";
+import { Home, LayoutGrid, MessageSquare, Plus } from "lucide-react";
 import { useState } from "react";
+import { TOOLS_PANELS, type MobileToolsPanel } from "../../shared/components/mobile-shell-actions";
 import { useChatStore } from "../../shared/stores/chat.store";
 import { useUIStore } from "../../shared/stores/ui.store";
 import { cn } from "../../shared/lib/utils";
@@ -9,17 +10,6 @@ const MODE_OPTIONS = [
   { mode: "roleplay" as const, label: "Roleplay", color: "text-pink-400", border: "border-pink-500/30", gradient: "from-pink-500/20 to-pink-500/5" },
   { mode: "game" as const, label: "Game", color: "text-amber-400", border: "border-amber-500/30", gradient: "from-amber-500/20 to-amber-500/5" },
 ];
-
-export const TOOLS_PANELS = [
-  { panel: "bot-browser" as const, icon: Bot, label: "Browser", gradient: "from-cyan-500 to-blue-500", color: "text-cyan-400" },
-  { panel: "characters" as const, icon: Users, label: "Characters", gradient: "from-pink-500 to-rose-500", color: "text-rose-400" },
-  { panel: "lorebooks" as const, icon: BookOpen, label: "Lorebooks", gradient: "from-amber-500 to-orange-500", color: "text-amber-400" },
-  { panel: "presets" as const, icon: FileText, label: "Presets", gradient: "from-purple-500 to-violet-500", color: "text-violet-400" },
-  { panel: "connections" as const, icon: Link, label: "Connections", gradient: "from-sky-500 to-blue-500", color: "text-sky-400" },
-  { panel: "agents" as const, icon: Sparkles, label: "Agents", gradient: "from-pink-500 to-purple-500", color: "text-pink-400" },
-  { panel: "personas" as const, icon: User, label: "Personas", gradient: "from-emerald-500 to-teal-500", color: "text-emerald-400" },
-  { panel: "settings" as const, icon: Settings, label: "Settings", gradient: "from-zinc-400 to-zinc-500", color: "text-zinc-300" },
-] as const;
 
 export function MobileTabBar({
   professorMariOpen,
@@ -74,7 +64,7 @@ export function MobileTabBar({
     }
   };
 
-  const openPanel = (panel: typeof TOOLS_PANELS[number]["panel"]) => {
+  const openPanel = (panel: MobileToolsPanel) => {
     const wasThisPanel = rightPanelOpen && rightPanel === panel;
     closeAll();
     if (!wasThisPanel) openRightPanel(panel);
